refactor(website): move directory `routes` to `views`
