+++ website/src/components/rsv/RsvSingleVariantPage.astro
@@ -5,11 +5,11 @@ import LocationTimeFilter from '../../components/LocationTimeFilter.astro';
 import VariantFilter from '../../components/VariantFilter.astro';
 import { getLapisUrl } from '../../config';
 import BaseLayout from '../../layouts/base/BaseLayout.astro';
-import { defaultTablePageSize } from '../../routes/View';
-import { chooseGranularityBasedOnDateRange, getLapisMutations, getLocationSubdivision } from '../../routes/helpers';
-import type { OrganismViewKey, singleVariantViewKey } from '../../routes/routing';
-import { ServerSide } from '../../routes/serverSideRouting';
 import { organismConfig, type Organisms } from '../../types/Organism';
+import { defaultTablePageSize } from '../../views/View';
+import { chooseGranularityBasedOnDateRange, getLapisMutations, getLocationSubdivision } from '../../views/helpers';
+import type { OrganismViewKey, singleVariantViewKey } from '../../views/routing';
+import { ServerSide } from '../../views/serverSideRouting';
 
 interface Props {
     organismViewKey: Extract<
