+++ website/src/pages/covid/compare-side-by-side.astro
@@ -3,11 +3,11 @@ import ComponentWrapper from '../../components/ComponentWrapper.astro';
 import JointFilter from '../../components/covidCompareVariantsView/JointFilter.astro';
 import { getLapisUrl } from '../../config';
 import BaseLayout from '../../layouts/base/BaseLayout.astro';
-import { defaultTablePageSize } from '../../routes/View';
-import { chooseGranularityBasedOnDateRange, getLapisMutations } from '../../routes/helpers';
-import type { OrganismViewKey } from '../../routes/routing';
-import { ServerSide } from '../../routes/serverSideRouting';
 import { Organisms } from '../../types/Organism';
+import { defaultTablePageSize } from '../../views/View';
+import { chooseGranularityBasedOnDateRange, getLapisMutations } from '../../views/helpers';
+import type { OrganismViewKey } from '../../views/routing';
+import { ServerSide } from '../../views/serverSideRouting';
 
 const organismViewKey: OrganismViewKey = 'covid.compareVariantsView';
 const view = ServerSide.routing.getOrganismView(organismViewKey);
