+++ website/src/pages/covid/single-variant.astro
@@ -6,16 +6,16 @@ import VariantFilter from '../../components/VariantFilter.astro';
 import { CollectionsList } from '../../components/covidAnalyzeSingleVariantView/CollectionsList';
 import { getDashboardsConfig, getLapisUrl } from '../../config';
 import BaseLayout from '../../layouts/base/BaseLayout.astro';
-import { defaultTablePageSize } from '../../routes/View';
+import { Organisms } from '../../types/Organism';
+import { defaultTablePageSize } from '../../views/View';
 import {
     chooseGranularityBasedOnDateRange,
     getLapisMutations,
     getLocationSubdivision,
     lineageKey,
-} from '../../routes/helpers';
-import { type OrganismViewKey } from '../../routes/routing';
-import { ServerSide } from '../../routes/serverSideRouting';
-import { Organisms } from '../../types/Organism';
+} from '../../views/helpers';
+import { type OrganismViewKey } from '../../views/routing';
+import { ServerSide } from '../../views/serverSideRouting';
 
 const organismViewKey: OrganismViewKey = 'covid.singleVariantView';
 const view = ServerSide.routing.getOrganismView(organismViewKey);
