+++ website/src/pages/flu/h5n1/index.astro
@@ -1,5 +1,5 @@
 ---
-import { ServerSide } from '../../../routes/serverSideRouting';
+import { ServerSide } from '../../../views/serverSideRouting';
 
 return Astro.redirect(ServerSide.routing.getOrganismView('h5n1.singleVariantView').getDefaultPageState());
 ---
