+++ website/src/pages/flu/h5n1/sequencing-efforts.astro
@@ -3,11 +3,11 @@ import ComponentsGrid from '../../../components/ComponentsGrid.astro';
 import LocationTimeFilter from '../../../components/LocationTimeFilter.astro';
 import { getLapisUrl } from '../../../config';
 import BaseLayout from '../../../layouts/base/BaseLayout.astro';
-import { defaultTablePageSize } from '../../../routes/View';
-import { chooseGranularityBasedOnDateRange, getLocationSubdivision } from '../../../routes/helpers';
-import type { OrganismViewKey } from '../../../routes/routing';
-import { ServerSide } from '../../../routes/serverSideRouting';
 import { Organisms } from '../../../types/Organism';
+import { defaultTablePageSize } from '../../../views/View';
+import { chooseGranularityBasedOnDateRange, getLocationSubdivision } from '../../../views/helpers';
+import type { OrganismViewKey } from '../../../views/routing';
+import { ServerSide } from '../../../views/serverSideRouting';
 
 const organismViewKey: OrganismViewKey = 'h5n1.sequencingEffortsView';
 const view = ServerSide.routing.getOrganismView(organismViewKey);
