+++ website/src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import BaseLayout from '../layouts/base/BaseLayout.astro';
-import { ServerSide } from '../routes/serverSideRouting';
 import { organismConfig } from '../types/Organism';
+import { ServerSide } from '../views/serverSideRouting';
 ---
 
 <style>
