+++ website/src/pages/mpox/single-variant.astro
@@ -5,17 +5,17 @@ import LocationTimeFilter from '../../components/LocationTimeFilter.astro';
 import VariantFilter from '../../components/VariantFilter.astro';
 import { getLapisUrl } from '../../config';
 import BaseLayout from '../../layouts/base/BaseLayout.astro';
-import { defaultTablePageSize } from '../../routes/View';
+import { Organisms } from '../../types/Organism';
+import { defaultTablePageSize } from '../../views/View';
 import {
     chooseGranularityBasedOnDateRange,
     cladeKey,
     getLapisMutations,
     getLocationSubdivision,
     lineageKey,
-} from '../../routes/helpers';
-import type { OrganismViewKey } from '../../routes/routing';
-import { ServerSide } from '../../routes/serverSideRouting';
-import { Organisms } from '../../types/Organism';
+} from '../../views/helpers';
+import type { OrganismViewKey } from '../../views/routing';
+import { ServerSide } from '../../views/serverSideRouting';
 
 const organismViewKey: OrganismViewKey = 'mpox.singleVariantView';
 const view = ServerSide.routing.getOrganismView(organismViewKey);
