+++ website/src/pages/rsv-a/single-variant.astro
@@ -1,7 +1,7 @@
 ---
 import RsvSingleVariantPage from '../../components/rsv/RsvSingleVariantPage.astro';
-import { singleVariantViewKey } from '../../routes/routing';
 import { Organisms } from '../../types/Organism';
+import { singleVariantViewKey } from '../../views/routing';
 ---
 
 <RsvSingleVariantPage organismViewKey={`${Organisms.rsvB}.${singleVariantViewKey}`} />
