+++ website/src/pages/rsv-b/index.astro
@@ -1,4 +1,4 @@
 ---
-import { ServerSide } from '../../routes/serverSideRouting';
+import { ServerSide } from '../../views/serverSideRouting';
 return Astro.redirect(ServerSide.routing.getOrganismView('rsvB.singleVariantView').getDefaultPageState());
 ---
