+++ website/src/pages/rsv-b/sequencing-efforts.astro
@@ -1,7 +1,7 @@
 ---
 import RsvSequencingEffortsPage from '../../components/rsv/RsvSequencingEffortsPage.astro';
-import { sequencingEffortsViewKey } from '../../routes/routing';
 import { Organisms } from '../../types/Organism';
+import { sequencingEffortsViewKey } from '../../views/routing';
 ---
 
 <RsvSequencingEffortsPage organismViewKey={`${Organisms.rsvB}.${sequencingEffortsViewKey}`} />
