+++ website/src/pages/west-nile/index.astro
@@ -1,5 +1,5 @@
 ---
-import { ServerSide } from '../../routes/serverSideRouting';
+import { ServerSide } from '../../views/serverSideRouting';
 
 return Astro.redirect(ServerSide.routing.getOrganismView('westNile.singleVariantView').getDefaultPageState());
 ---
