+++ website/src/views/covid.ts
