+++ website/src/views/mpox.ts
